+++ Models/ProductAttributeValue.cs
@@ -11,7 +11,7 @@ namespace Nwazet.Commerce.Models {
         [Required, DisplayName("Value Name")]
         public string Text { get; set; }
         [Required, DisplayName("Price Adjustment")]
-        public double PriceAdjustment { get; set; }
+        public decimal PriceAdjustment { get; set; }
         [DefaultValue(false), DisplayName("Is Line Adjustment")]
         public bool IsLineAdjustment { get; set; }
         [DefaultValue(0), DisplayName("Sort Order")]
@@ -26,7 +26,7 @@ namespace Nwazet.Commerce.Models {
                         var attrSettings = av[1].Split(',');
                         return new ProductAttributeValue {
                             Text = av[0],
-                            PriceAdjustment = Convert.ToDouble(attrSettings[0]),
+                            PriceAdjustment = Convert.ToDecimal(attrSettings[0]),
                             IsLineAdjustment = Convert.ToBoolean(attrSettings[1]),
                             // Check if sort order value is present, didn't exist in previous versions
                             SortOrder = attrSettings.Length > 2 ? Convert.ToInt32(attrSettings[2]) : 0,
