+++ Models/ProductPartRecord.cs
@@ -8,11 +8,11 @@ namespace Nwazet.Commerce.Models {
             ShippingCost = null;
         }
         public virtual string Sku { get; set; }
-        public virtual double Price { get; set; }
-        public virtual double DiscountPrice { get; set; }
+        public virtual decimal Price { get; set; }
+        public virtual decimal DiscountPrice { get; set; }
         public virtual bool IsDigital { get; set; }
         public virtual bool ConsiderInventory { get; set; } //applies for digital products,m telling whether to consider a limited inventory
-        public virtual double? ShippingCost { get; set; }
+        public virtual decimal? ShippingCost { get; set; }
         public virtual double Weight { get; set; }
         public virtual string Size { get; set; }
         public virtual int Inventory { get; set; }
