+++ Models/WeightBasedShippingMethodPart.cs
@@ -20,7 +20,7 @@ namespace Nwazet.Commerce.Models {
             set { Store(r => r.ShippingCompany, value); }
         }
 
-        public double Price {
+        public decimal Price {
             get { return Retrieve(r => r.Price); }
             set { Store(r => r.Price, value); }
         }
@@ -56,7 +56,7 @@ namespace Nwazet.Commerce.Models {
             var fixedCost = quantities
                 .Where(pq => pq.Product.ShippingCost != null && pq.Product.ShippingCost >= 0 && !pq.Product.IsDigital)
 // ReSharper disable PossibleInvalidOperationException
-                .Sum(pq => pq.Quantity*(double) pq.Product.ShippingCost);
+                .Sum(pq => pq.Quantity*(decimal) pq.Product.ShippingCost);
 // ReSharper restore PossibleInvalidOperationException
             var weight = quantities
                 .Where(pq => (pq.Product.ShippingCost == null || pq.Product.ShippingCost < 0) && !pq.Product.IsDigital)
@@ -69,7 +69,7 @@ namespace Nwazet.Commerce.Models {
             }
         }
 
-        private ShippingOption GetOption(double price) {
+        private ShippingOption GetOption(decimal price) {
             return new ShippingOption {
                 Description = Name,
                 Price = price,
